+++ NEWS.md
@@ -1,4 +1,6 @@
 # Changes in version 0.4.1.9000
+##New Functionality
+* New sampler type `rw-ref`, an implementation of the reflecting random walk described in Yang and Rodriguez 2013. The sampler reflects on the bounds of the associated prior distribution.
 ##Changed functionality
 * runtime plotting during the MCMC procedure `de_mcmc(... , plot=TRUE)` now omits fixed parameters
 * correlation coefs in pairs plot now scale with strength of correlation